try langchain OpenAI adapter; fix enums
