+++ .gitignore
@@ -1,0 +1,2 @@
+__pycahe__
+*.pyc
